pyproject.toml
[tool.poetry]
name = "myuserbot"
version = "0.1.0"
description = ""
authors = ["XetPy1030 <xetpy1030@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.38.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.22.0"}
loguru = "^0.7.2"
celery = {extras = ["redis"], version = "^5.4.0"}
arrow = "^1.3.0"
aioredis = "^2.0.1"
asgiref = "^3.8.1"
python-dotenv = "^1.0.1"
aerich = {extras = ["toml"], version = "^0.8.0"}


[tool.aerich]
tortoise_orm = "app.database.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
